+++ carta/cpp/core/Data/Image/LayerGroup.cpp
@@ -525,7 +525,7 @@ RegionHistogramData LayerGroup::_getPixels2Histogram(int frameLow, int frameHigh
     return results;
 }
 
-std::vector<float> LayerGroup::_getRasterImageData(double xMin, double xMax, double yMin, double yMax,
+std::vector<float> LayerGroup::_getRasterImageData(int xMin, int xMax, int yMin, int yMax,
     int mip, double minIntensity, int frameLow, int frameHigh, int stokeFrame) const {
     std::vector<float> results;
     int dataIndex = _getIndexCurrent();
